+++ src/image_filter.h
@@ -28,8 +28,21 @@ extern "C" {
 #endif
 
 #include "types.h"
-#include "photorec.h"
+#include <stdint.h>
 
+struct image_size_filter_struct
+{
+  uint64_t min_file_size;  /* 0 = no limit */
+  uint64_t max_file_size;  /* 0 = no limit */
+  uint32_t min_width;      /* 0 = no limit */
+  uint32_t max_width;      /* 0 = no limit */
+  uint32_t min_height;     /* 0 = no limit */
+  uint32_t max_height;     /* 0 = no limit */
+  uint64_t min_pixels;     /* 0 = no limit (width × height) */
+  uint64_t max_pixels;     /* 0 = no limit (width × height) */
+};
+
+typedef struct image_size_filter_struct image_size_filter_t;
 
 int has_any_image_size_filter(const image_size_filter_t *filter);
 int should_skip_image_by_dimensions(const image_size_filter_t *filter, uint32_t width, uint32_t height);
